requirements.txt
# Core ML/DL dependencies
torch>=2.0.0,<2.3.0
torchvision>=0.15.0,<0.18.0
pytorch-lightning>=2.0.0,<2.2.0

# Diffusion model dependencies
diffusers>=0.24.0,<0.28.0
accelerate>=0.24.0,<0.28.0
transformers>=4.30.0,<4.40.0

# Configuration and experiment tracking
hydra-core>=1.3.0,<1.4.0
wandb>=0.15.0,<0.17.0

# Scientific computing
numpy>=1.24.0,<1.27.0
scipy>=1.10.0,<1.13.0
scikit-image>=0.21.0,<0.23.0

# Computer vision and image processing
kornia>=0.7.0,<0.8.0
tifffile>=2023.0.0
Pillow>=9.4.0,<11.0.0

# Data storage and compression
zarr>=2.16.0,<3.0.0
numcodecs>=0.12.1,<0.13.0

# Biological image analysis
cellpose>=2.0.0,<3.1.0

# Visualization
matplotlib>=3.7.0,<3.9.0
seaborn>=0.12.0,<0.14.0

# Utilities
tqdm>=4.65.0,<5.0.0
einops>=0.6.0,<0.8.0
requests>=2.31.0,<3.0.0

# Development tools
pytest>=7.3.0,<8.2.0
black>=23.0.0,<25.0.0
isort>=5.12.0,<6.0.0
flake8>=6.0.0,<8.0.0

# Jupyter environment
jupyter>=1.0.0,<2.0.0
ipywidgets>=8.0.0,<9.0.0

# Optional performance dependencies
psutil>=5.9.0,<6.0.0  # For memory monitoring
rawpy>=0.18.0,<0.20.0  # For RAW image processing (SID dataset)


